pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "a2a-mcp-contractor-automation"
version = "0.1.0"
description = "A2A MCP Contractor Automation System"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Proprietary"}
authors = [
    {name = "A2A Team", email = "info@a2a.example.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-dotenv>=0.19.0",
    "aiohttp>=3.8.0",
    "httpx>=0.23.0",
    "pydantic>=1.10.0",
    "chromadb>=0.3.0",
    "a2a-sdk>=1.0.0",  # Replace with actual A2A SDK package name
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.910",
    "flake8>=4.0.0",
]

[project.scripts]
a2a-contractor = "agents.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "config*", "models*", "utils*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
